+++ capi/autd3capi-wrapper-generator/src/python.rs
@@ -4,7 +4,7 @@
  * Created Date: 25/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 29/11/2023
+ * Last Modified: 01/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -223,6 +223,8 @@ impl Generator for PythonGenerator {
                     | "DynamicDatagramPack"
                     | "DynamicDatagramPack2"
                     | "DynamicLinkBuilder"
+                    | "DynamicConfigureDebugOutputIdx"
+                    | "DynamicDebugOutIdxOp"
             )
         }));
         self