+++ src/autd3/src/prelude.rs
@@ -4,7 +4,7 @@
  * Created Date: 27/04/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 22/11/2023
+ * Last Modified: 01/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022-2023 Shun Suzuki. All rights reserved.
@@ -28,7 +28,7 @@ pub use autd3_driver::{
     autd3_device::AUTD3,
     common::{Drive, EmitIntensity, SamplingConfiguration},
     datagram::{
-        Clear, ConfigureDebugOutoutIdx, ConfigureModDelay, DatagramT, FocusSTM, GainSTM,
+        Clear, ConfigureDebugOutputIdx, ConfigureModDelay, DatagramT, FocusSTM, GainSTM,
         Modulation, ModulationProperty, Silencer, Stop, Synchronize, UpdateFlags,
     },
     defined::{float, METER, MILLIMETER, PI},